照会画面の遷移のリファクタリング

+++ src/main/java/jp/lg/asp/accommodation/dto/FukaTaxDetailDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class FukaTaxDetailDto {
 	private String label;
-	private Long taxRate;
+	private BigDecimal taxRate;
 
 	// 💡 修正：個別の行は 0 や未入力（null）を許容するため制約を削除
 	// 未入力時は Service 層で 0 に補完するぜ
